pull `secureStorage` out into its own file
